### 2020-12-16 #2913 @whedon
[PRE REVIEW]: NLSIG-COVID19Lab
**Submitting author:** @somefunagba (<a href="http://orcid.org/0000-0002-5171-8026">Oluwasegun Somefun</a>)
**Repository:** <a href="https://github.com/somefunAgba/NLSIG-COVID19Lab" target ="_blank">https://github.com/somefunAgba/NLSIG-COVID19Lab</a>
**Version:** v1.1.0
