### 2020-12-16 #2913 @somefunAgba
Hello. Thanks @kthyng.

All code in the repository apart from those in the helpers directory was written by me.

-- helpers directory
The code in the helpers directory are also open-source matlab functions.
1.  LinePlotReducer.m and reduce_plot.m is an alternative to matlab's plot function.
- Status: unused, added for possible use.

2. SLMtools is a custom matlab toolbox for fitting splines on data.
- Status: minimal usage. This was used to smooth the noisy data, so as to reduce the time spent in finding zero-crossings.
- active only when, bootstrapping action is specified by the user.

-- main directories
1. cvmain
- contains the main functions for the NLSIG-COVID19Lab
2. applet
- contains the MATLAB GUI Application file for NLSIG-COVID19Lab
3. local
- local database directory for reference
4. data
-  actual database directory
5.  ui
- contains user application scripts for executing NLSIG-COVID19Lab
6. assets, measures
- dynamically created directories: respectively:
- stores plots of the  infections and deaths growth model
- stores the model metrics for the COVID-19 phenomena in each locale successfully queried.

7. osspaper
- directory for paper.md submitted to JOSS.
