### 2020-12-17 e04a759
Updated stats
feat: KS statistics for goodness of fit
