+++ cvmain/nlsigfit/nsligfp_dkw_applet.m
@@ -1,9 +1,6 @@
 function [y_sollb,dy_sollb,sol_lb,y_solub,dy_solub,sol_ub,...
     fitstatslb,fitstatsub] ...
-    = nsligfp_dkw_applet(app,x_data,y_data,dy_data,...
-    sol,y_sol,dy_sol,len_sol,...
-    imposeconstr,chngsolver,fitstats, ... 
-    y_mdlfun,dy_dx_mdlfun,nlsigprob,x0,n_ips,newoptins)
+    = nsligfp_dkw_applet(app,x_data,y_sol,dy_sol,fitstats,fitopts,sbounds,lubnds)
 %NLSIGFP_BOOTSTRAP Compute uncertainty on data to nlsig fit/prediction
 
 % display bootstrapping progress
@@ -16,68 +13,46 @@ yref = y_sol;
 % yref = y_data;
 dyref = dy_sol;
 
-y_lb = yref - fitstats.ciE;
-dy_lb = dyref - fitstats.dciE;
-bt_objsse = sum((y_mdlfun - y_lb).^2) + sum((dy_dx_mdlfun - dy_lb).^2);
-nlsigprob.Objective = bt_objsse;
-x0lb = x0;
-x0lb.ymin = x0.ymin - fitstats.ciE;
-x0lb.ymax = x0.ymax - fitstats.ciE;
-sol_lb = fitnlsig(nlsigprob,x0lb,imposeconstr,chngsolver,newoptins);
-
-y_ub = yref + fitstats.ciE;
-dy_ub = dyref + fitstats.dciE;
-bt_objsse = sum((y_mdlfun - y_ub).^2) + sum((dy_dx_mdlfun - dy_ub).^2);
-nlsigprob.Objective = bt_objsse;
-x0ub = x0;
-x0ub.ymin = x0.ymin + fitstats.ciE;
-x0ub.ymax = x0.ymax + fitstats.ciE;
-sol_ub = fitnlsig(nlsigprob,x0ub,imposeconstr,chngsolver,newoptins);
-
-app.StatusLabel.Text =  " Done.";
-app.StatusLabel.FontColor = boldgreen;
-
-% LB_CI
-y_estlb = evaluate(y_mdlfun,sol_lb);
-dy_dx_estlb = evaluate(dy_dx_mdlfun, sol_lb);
-fitstatslb = calc_stats(y_data,y_estlb,dy_data,dy_dx_estlb,len_sol,n_ips);
-
-% UB_CI
-y_estub = evaluate(y_mdlfun,sol_ub);
-dy_dx_estub = evaluate(dy_dx_mdlfun, sol_ub);
-fitstatsub = calc_stats(y_data,y_estub,dy_data,dy_dx_estub,len_sol,n_ips);
-
-%% Augment un-optimized parameters with the optimized solutions struct
 
 % LB
-sol_lb.check_constraints = 0;
-sol_lb.n = n_ips;
-sol_lb.shape = 's';
+y_lb = yref - fitstats.ciE;
+dy_lb = dyref - fitstats.dciE;
+% bt_objsse = sum((y_mdlfun - y_lb).^2) + sum((dy_dx_mdlfun - dy_lb).^2);
+% nlsigprob.Objective = bt_objsse;
+
+% Set up fit options structure
+fitoptslb = fitopts;
+fitoptslb.ymin = fitopts.ymin - fitstats.ciE;
+fitoptslb.ymin(1) = fitopts.ymin(1);
+fitoptslb.ymax = fitopts.ymax - fitstats.ciE;
+% Fit
+[sol_lb,~,~,~,fitstatslb] = ...
[…]
