+++ cvmain/cvcore/cov19_nlsigquery.m
@@ -138,9 +138,8 @@ if boots == true
 else
 [gy_sollb,gdy_sollb,gsol_lb,gy_solub,gdy_solub,gsol_ub,...
     gbest_fitstatslb,gbest_fitstatsub] ...
-    = nsligfp_dkw(x_data,y_data,dy_data,gsol,gy_sol,gdy_sol,len_sol,...
-    imposeconstr,chngsolver,gbest_fitstats,...
-    gy_mdlfun,gdy_dx_mdlfun,gnlsigprob,gx0,gn_ips,gnewoptins); %#ok<ASGLU>
+    = nsligfp_dkw(x_data,gy_sol,gdy_sol,fitstats,...
+    fitopts,startingbounds,lubnds); %#ok<ASGLU>
 end
 
 %% Second-Pass 
@@ -174,7 +173,7 @@ if finer == true
         y_mdlfun,dy_dx_mdlfun,nlsigprob,x0,n_ips,gnewoptins,fitopts] ...
         = nsligfp(x_data,sy_data,sdy_data,ig_opts,startingbounds,lubnds,...
         len_sol,imposeconstr,chngsolver); %#ok<ASGLU>
-
+    
     if boots == true
         [y_sollb,dy_sollb,sol_lb,y_solub,dy_solub,sol_ub,...
             best_fitstatslb,best_fitstatsub] ...
@@ -184,10 +183,8 @@ if finer == true
     else
         [y_sollb,dy_sollb,sol_lb,y_solub,dy_solub,sol_ub,...
             best_fitstatslb,best_fitstatsub] ...
-            = nsligfp_dkw(x_data,sy_data,sdy_data,sol,y_sol,dy_sol,len_sol,...
-            imposeconstr,chngsolver,gbest_fitstats,...
-            y_mdlfun,dy_dx_mdlfun,nlsigprob,x0,n_ips,gnewoptins); %#ok<ASGLU>
-        
+            = nsligfp_dkw(x_data,y_sol,dy_sol,gbest_fitstats,...
+            fitopts,startingbounds,lubnds); %#ok<ASGLU>
     end
 end
 
