+++ cvmain/cvcore/cov19_nlsigquery_applet.m
@@ -140,9 +140,8 @@ if boots == true
 else
 [gy_sollb,gdy_sollb,gsol_lb,gy_solub,gdy_solub,gsol_ub,...
     gbest_fitstatslb,gbest_fitstatsub] ...
-    = nsligfp_dkw_applet(app,x_data,y_data,dy_data,gsol,gy_sol,gdy_sol,len_sol,...
-    imposeconstr,chngsolver,gbest_fitstats,...
-    gy_mdlfun,gdy_dx_mdlfun,gnlsigprob,gx0,gn_ips,gnewoptins); %#ok<ASGLU>
+    = nsligfp_dkw_applet(app,x_data,gy_sol,gdy_sol,fitstats,...
+    fitopts,startingbounds,lubnds); %#ok<ASGLU>
 end
 
 
@@ -187,9 +186,8 @@ if finer == true
     else
         [y_sollb,dy_sollb,sol_lb,y_solub,dy_solub,sol_ub,...
             best_fitstatslb,best_fitstatsub] ...
-            = nsligfp_dkw_applet(app,x_data,sy_data,sdy_data,sol,y_sol,dy_sol,len_sol,...
-            imposeconstr,chngsolver,gbest_fitstats,...
-            y_mdlfun,dy_dx_mdlfun,nlsigprob,x0,n_ips,gnewoptins); %#ok<ASGLU>
+            = nsligfp_dkw_applet(app,x_data,y_sol,dy_sol,gbest_fitstats,...
+            fitopts,startingbounds,lubnds); %#ok<ASGLU>
         
     end
 end
