+++ cvmain/nlsigcore/nlsig.m
@@ -1,6 +1,6 @@
-function [y,dy_dx] = nlsig(x,eq,iopts)
+function [y,dy_dx,d2y_dx2,jacob_iy_d,jacoby,hessy] = nlsig(x,eq,iopts)
 %nLOGISTIC-SIGMOID function
-% logistic-sigmoid function for 
+% logistic-sigmoid function for
 % multiple peak inflection points, i = 1, ..., n.
 % $$ y = f(x) $$
 %
@@ -8,14 +8,14 @@ function [y,dy_dx] = nlsig(x,eq,iopts)
 %
 % [y,dy_dx] = nlsig(x,eq,iopts)
 %
-% Inputs: 
-% x: scalar | vector 
+% Inputs:
+% x: scalar | vector
 % eq: 0 or 1
 % iopts: struct,
 %
 % if eq = 1 : all struct elements are scalar
 % else :
-% scalar size: 1 by 1 
+% scalar size: 1 by 1
 % iopts.n = 1;
 % iopts.shape = 's';
 %
@@ -27,7 +27,7 @@ function [y,dy_dx] = nlsig(x,eq,iopts)
 % iopts.xmin
 % iopts.ymax
 % iopts.ymin
-% 
+%
 % Outputs:
 % y: scalar | vector
 % dy_dx: scalar | vector
@@ -36,501 +36,677 @@ function [y,dy_dx] = nlsig(x,eq,iopts)
 % |oasomefun@futa.edu.ng| 2020.
 
 if eq == 1
-    [y,dy_dx] = nlsig_eq(x,iopts);
+    if nargout == 1
+        y = nlsig_eq(x,iopts);
+    elseif nargout == 2
+        [y,dy_dx] = nlsig_eq(x,iopts);
+    elseif nargout == 3
+        [y,dy_dx,d2y_dx2] = nlsig_eq(x,iopts);
+    elseif nargout == 4
+        [y,dy_dx,d2y_dx2,jacob_iy_d] = nlsig_eq(x,iopts);
+    elseif nargout == 5
+        [y,dy_dx,d2y_dx2,jacob_iy_d,jacoby] = nlsig_eq(x,iopts);
+    elseif nargout == 6
+        [y,dy_dx,d2y_dx2,jacob_iy_d,jacoby,hessy] = nlsig_eq(x,iopts);
+    end
 elseif eq == 0
-    [y,dy_dx] = nlsig(x,iopts);
+    if nargout == 1
+        y = nlsig(x,iopts);
+    elseif nargout == 2
+        [y,dy_dx] = nlsig(x,iopts);
+    elseif nargout == 3
+        [y,dy_dx,d2y_dx2] = nlsig(x,iopts);
+    elseif nargout == 4
+        [y,dy_dx,d2y_dx2,jacob_iy_d] = nlsig(x,iopts);
+    elseif nargout == 5
+        [y,dy_dx,d2y_dx2,jacob_iy_d,jacoby] = nlsig(x,iopts);
+    elseif nargout == 6
+        [y,dy_dx,d2y_dx2,jacob_iy_d,jacoby,hessy] = nlsig(x,iopts);
+    end
 else
     err_msg = sprintf("Try again! syntax: [y,dy_dx] = nlsig(x,eq,iopts),\n"+...
-                    "Note: ensure eq is either 0 or 1!");
+        "Note: ensure eq is either 0 or 1!");
     error(err_msg);
 end
 
 %% general (equal|unequal) interval
-function [y,dy_dx] = nlsig(x,iopts)
-%nLOGISTIC-SIGMOID
-% logistic-sigmoid function for 
-% multiple inflection points, i = 1, ..., n.
-% arbitrarily divided intervals.
-%
[…]
