+++ cvmain/nlsigfit/set_probopts.m
@@ -163,7 +163,7 @@ end
 % create objective function as an optimization expression
 
 if len_sol == 6
-    objfun = @(lambda,xmax,xmin,ymax,ymin,xpks,xend,yend)...
+    objfun = @(lambda,xmax,xmin,ymax,ymin,xpks)...
         nlsigeval(x_data,n_ips,shape,base,lambda,xmax,xmin,ymax,ymin,xpks);
     [y_mdlfun, dy_dx_mdlfun] = fcn2optimexpr(objfun,lambda,xmax,xmin,ymax,ymin,xpks,...
         'ReuseEvaluation',false,...
