+++ cvmain/nlsigfit/nlsigopt_test.m
@@ -70,9 +70,8 @@ if boots == true
 else
 [y_sollb,dy_sollb,sol_lb,y_solub,dy_solub,sol_ub,...
     best_fitstatslb,best_fitstatsub] ...
-    = nsligfp_dkw(x_data,y_data,dy_data,sol,y_sol,dy_sol,len_sol,...
-    imposeconstr,chngsolver,fitstats,... 
-    y_mdlfun,dy_dx_mdlfun,nlsigprob,x0,n_ips,newoptins);
+    = nsligfp_dkw(x_data,y_sol,dy_sol,fitstats,...
+    nlsigfit_opts,sbounds,lubnds);
 end
 %% 
 % Predict using 'sol', the obtained 'best'-fitting solution 
