+++ osspaper/AVR_ASM/cmake-build-debug/CMakeFiles/clion-log.txt
@@ -1,0 +1,1 @@
+CMakeLists.txt not found in D:\AVR_ASM Select CMakeLists.txt file...
