### 2021-01-01 2a0ea8a
Delete CmakeListsAVR.cmake
