+++ osspaper/AVR_ASM/CmakeListsAVR.cmake
@@ -1,94 +1,0 @@
-# I target a recent cmake, it shouldn't be a problem on a dev machine
-cmake_minimum_required(VERSION 3.11)
-# Project name
-project("My Firmware")
-
-# Product filename
-set(PRODUCT_NAME my_firmware)
-
-## AVR Chip Configuration
-# 8Mhz, this should match the crystal on your board,
-# I use 8Mhz and 3.3V for the lowest power consumption
-set(F_CPU 8000000UL)
-# CPU, you can find the list here:
-# https://gcc.gnu.org/onlinedocs/gcc/AVR-Options.html
-set(MCU atmega328p)
-# Default Baudrate for UART, read avr include/util/setbaud.h for usage
-set(BAUD 9600)
-# The programmer to use, read avrdude manual for list
-set(PROG_TYPE avrispmkII)
-
-# AVR Fuses, must be in concordance with your hardware and F_CPU
-# http://eleccelerator.com/fusecalc/fusecalc.php?chip=atmega328p
-set(E_FUSE 0xfd)
-set(H_FUSE 0xda)
-set(L_FUSE 0xfd)
-set(LOCK_BIT 0xff)
-
-
-# Use AVR GCC toolchain
-set(CMAKE_SYSTEM_NAME Generic)
-set(CMAKE_CXX_COMPILER avr-g++)
-set(CMAKE_C_COMPILER avr-gcc)
-set(CMAKE_ASM_COMPILER avr-gcc)
-
-
-# Pass defines to compiler
-add_definitions(
-    -DF_CPU=${F_CPU}
-    -DBAUD=${BAUD}
-)
-# mmcu MUST be passed to bot the compiler and linker, this handle the linker
-set(CMAKE_EXE_LINKER_FLAGS -mmcu=${MCU})
-
-add_compile_options(
-    -mmcu=${MCU} # MCU
-    -std=gnu99 # C99 standard
-    -Os # optimize
-    -Wall # enable warnings
-    -Wno-main
-    -Wundef
-    -pedantic
-    -Wstrict-prototypes
-    -Werror
-    -Wfatal-errors
-    -Wl,--relax,--gc-sections
-    -g
-    -gdwarf-2
-    -funsigned-char # a few optimizations
-    -funsigned-bitfields
-    -fpack-struct
-    -fshort-enums
-    -ffunction-sections
-    -fdata-sections
-    -fno-split-wide-types
-    -fno-tree-scev-cprop
-)
-file(GLOB SRC_FILES "src/*.c") # Load all files in src folder
-
-# Create one target
-add_executable(${PRODUCT_NAME} ${SRC_FILES})
-
-# Rename the output to .elf as we will create multiple files
-set_target_properties(${PRODUCT_NAME} PROPERTIES OUTPUT_NAME ${PRODUCT_NAME}.elf)
-
-# Strip binary for upload
-add_custom_target(strip ALL avr-strip ${PRODUCT_NAME}.elf DEPENDS ${PRODUCT_NAME})
-
-# Transform binary into hex file, we ignore the eeprom segments in the step
-add_custom_target(hex ALL avr-objcopy -R .eeprom -O ihex ${PRODUCT_NAME}.elf ${PRODUCT_NAME}.hex DEPENDS strip)
-# Transform binary into hex file, this is the eeprom part (empty if you don't
-# use eeprom static variables)
[…]
