+++ osspaper/AVR_ASM/avra_asm_makefile_tutorial/z_end.md
@@ -1,7 +1,0 @@
-You are encouraged to implement the simple Arduino project presented, download the files (there's a download button at the top) and build the project. If you're unfamiliar with the resistor values from the attached image, here's the handy cheat sheet I'm using.
-
-![Resistor cheat sheet](http://i.imgur.com/B2mK7PT.png)
-*Taken from the "Arduino Projects Book", licensed under CreativeCommons BY-NC-SA 3.0*
-
----
-Except where otherwise noted, this work is licensed under a <a rel="license" href="http://creativecommons.org/licenses/by/4.0/">Creative Commons Attribution 4.0 International License</a>.
