+++ osspaper/AVR_ASM/cmake-build-debug/CMakeFiles/TargetDirectories.txt
@@ -1,3 +1,0 @@
-D:/AVR_ASM/cmake-build-debug/CMakeFiles/AVR_ASM.dir
-D:/AVR_ASM/cmake-build-debug/CMakeFiles/edit_cache.dir
-D:/AVR_ASM/cmake-build-debug/CMakeFiles/rebuild_cache.dir
