### 2021-01-05 #2913 @will-rowe
Hi @kthyng

I'm sorry but with the latest lockdown just announced in the UK, I'm going to be looking after my kids with them home from primary school. I will be working as I can but my editorial response times will be much slower. I can still take this submission on with the understanding that I may be slow to handle this at times but if there is another editor with more capacity, that may be better for the submitter. I'm happy either way.
