### 2021-01-22 #2913 @majensen
Hello @nnadeau -- would you have some time to review [this work](#2913) for [JOSS](https://joss.theoj.org)? It would be great to have your MATLAB expertise. Please let me know in this thread or via maj -dot- fortinbras -at- gmail -dot- com. Thanks for your consideration! -- @majensen
