### 2021-01-22 #2913 @majensen
Hola @jsgalan - would you be willing to review [this work](#2913) for [JOSS](https://joss.theoj.org)? Hope to tap into your MATLAB expertise. Please let me know in this thread or at maj -dot- fortinbras -at- gmail -dot- com. I appreciate your consideration! --@majensen
