### 2021-01-22 #2913 @nnadeau
@majensen I no longer have access to a MATLAB instance, but if @somefunAgba can confirm it's compatible with Octave, then I can give it a proper review
