### 2021-01-22 #2913 @somefunAgba
@nnadeau @majensen
I have not tested it with Octave. It was all implemented in MATLAB2020a. However, I doubt Octave supports Matlab's problem-based optimization workflow which I heavily used in the current version of my code.
