### 2021-01-25 #2913 @nnadeau
@majensen if MATLAB is a requirement, then it would be best if someone else reviews it.

@somefunAgba, as a pre-review comment I would recommend updating the README with [a robust requirements section](https://joss.readthedocs.io/en/latest/review_checklist.html#functionality) such that reviews (and future users) know exactly what MATLAB version(s), toolboxes, and other dependencies are needed. This would avoid confusion and user experience impedance like @ManuelaS is facing.
