### 2021-01-26 #2913 @somefunAgba
@ManuelaS because the optimization workflow used in the package is a _problem-based_, as opposed to a _solver-based_ one, **R2017b** and above is needed.
No external dependencies needed, other than the presence of Matlab's
- Optimization Toolbox, and
- Statistics and Machine Learning Toolbox.
