### 2021-01-30 #2913 @majensen
@kakearney would you have time and interest to review [this work](#2913) for [JOSS](https://joss.theoj.org)? We could use your MATLAB expertise. Please let me know in this thread or at maj - dot - fortinbras - at - gmail - dot - com. I appreciate your consideration! --@majensen
