### 2021-02-01 #2913 @kakearney
I would only be able to review this work from a technical (i.e. Matlab) standpoint.  The topic itself is well outside my research focus, so I will have to accept the summary paper at face value, and assume the authors have properly described the state of the field, this entry's place in it, and that the chosen fitting techniques are appropriate.  If that is okay, I should be able to provide a review by the end of the month.
