### 2021-02-02 #3002 @whedon
[REVIEW]: NLSIG-COVID19Lab
**Submitting author:** @somefunagba (<a href="http://orcid.org/0000-0002-5171-8026">Oluwasegun Somefun</a>)
**Repository:** <a href="https://github.com/somefunAgba/NLSIG-COVID19Lab" target ="_blank">https://github.com/somefunAgba/NLSIG-COVID19Lab</a>
**Version:** 2.1.1
