### 2021-02-16 #3002 @agahkarakuzu
I gave this one a quick try (R2019b) after installing NLSIGCOVID19Lab as a Matlab app (applet), I could not find instructions regarding how to use GUI, so intuitively I tried clicking the `model` button:

```octave
Error using writetable (line 152)
Invalid parameter name: WriteMode.

Error in save_metrics (line 32)
    writetable(Tm, metricsfile,"Sheet",query_ccode,"WriteMode","overwritesheet");

Error in cov19_nlsigquery_applet (line 248)
save_metrics(focus,new_mts,country_code,time_data);

Error in nlsigcovid19lab/ModelButtonPushed (line 209)
                    cov19_nlsigquery_applet(search_code,update, ...

Error using matlab.ui.control.internal.controller.ComponentController/executeUserCallback (line 335)
Error while evaluating Button PrivateButtonPushedFcn.
```

Then I tried `Update Database` first, which threw me another error:
```octave
Error using upd_cdata_applet (line 294)

Oops! Something went wrong. Beyond my control.

Error in get_cdata_applet (line 40)
upd_cdata_applet(app, update, search_ccode);

Error in allcc_upd_applet (line 4)
[~,~,~] = get_cdata_applet(app,"ALL",1);

Error in nlsigcovid19lab/UpdateDatabaseButtonPushed (line 277)
            allcc_upd_applet(app);

Error using matlab.ui.control.internal.controller.ComponentController/executeUserCallback (line 335)
Error while evaluating Button PrivateButtonPushedFcn.
```
As the error message was not informative, I could not infer what to do next:
https://github.com/somefunAgba/NLSIG-COVID19Lab/blob/2a99f46ae19ae1b1433816228353fcb7739ce45f/cvmain/cvcore/upd_cdata_applet.m#L291-L292

I guess a good first step would be having a really simple "getting started with GUI" documentation. If there is, can you please point me to that?

I also noticed that the [tests](https://www.mathworks.com/help/matlab/matlab_prog/create-simple-test-suites.html) are missing, which is another important assessment criteria. For quite a while, hooking MATLAB CI tests to a GitHub repo was not an option, but now through Azure, it is easily [possible](https://marketplace.visualstudio.com/items?itemName=MathWorks.matlab-azure-devops-extension). There is a free license dedicated for providing public repos with CI. This repo uses the new app designer, which is a big advantage for testing GUI functionality. I think adding tests would be a second good step. I am happy to help out with Azure and code coverage integration when the tests are ready.
