### 2021-02-16 #3002 @somefunAgba
Thanks @agahkarakuzu. I just tested the applet twice on my end in R2020b, and no errors.
The applet was written in R2020a. There are internal catch blocks to ensure errors and exceptions will rarely occur and are rarely shown to the user, other things being equal.

Debugging the errors:
One.
```
Error using writetable (line 152)
Invalid parameter name: WriteMode.
```
SOLUTION:
I think the problem lies with string quotes which I used when writing  the code. It seems 2019b does not implicitly perform a conversion between arguments specified with character quotes, and those with string quotes.
**I will fix that to ensure compatibility with previous MATLAB versions.**

Two.

```
Error using upd_cdata_applet (line 294)
Oops! Something went wrong. Beyond my control.
```
SOLUTION:
I think the cause here, may also be what triggered the first error above. The production code is structured, in a to be in a way, such that any exception in the upd_cdata_applet file is returned as line 294.
**I will fix that as indicated above and also ensure the ME exceptions are printed also.**

**Other Issues:**
Thanks, I will create a separate file for the **getting started with GUI documentation** .
There is one in the README.md file, but given this, I will enhance it to be more useful

Finally, about the missing MATLAB unit tests, CI integrations through Azure. I will get myself acquainted as soon as possible.
I will inform you when these changes are made.
