### 2021-02-16 #3002 @agahkarakuzu
Thank you so much for the quick response @somefunAgba!

I think these initial simple fixes would give this codebase a good usability boost. I have no expertise on the scientific aspect of this work, but will do my best to improve user experience by playing the annoying user wherever possible to ensure that the expected outcomes find the user  :)

RE getting started docs, I would even briefly explain how to add your applet as an app to MATLAB as the first (or second) step, this convention is rather new and most matlab users are not accustomed to it.

Maybe using the wiki pages of your repo for documentation would be a good idea for the sake of content organization. Given that you already have figures etc uploaded, you can easily write up some informative markdowns.
