### 2021-02-18 #3002 @kakearney
I did some brief experimenting with this code, but so far have hit a few different problems that are preventing any in-depth analysis.

**Issue 1**

First, I encountered the same errors that @agahkarakuzu mentioned above when first opening the GUI and clicking the "Model" button without changing any of the default options.  It sounds like that problem was diagnosed, so I'll await an update with that fix.

** Issue 2**

My second test was to open the GUI and try to change the country code.  I changed the "Search Code" dropdown menu to "US" and received the following error:

```matlab
Error in get_cdata_applet (line 106)
    sheets = sheetnames(cbc_CV19datafiled);

Error in getcasesbycc_applet (line 15)
[T,status,ccs] = get_cdata_applet(app, country_code, update);

Error in nlsigcovid19lab/SearchCodesDropDownValueChanged (line 294)
            t = getcasesbycc_applet(app,string(value),0,'o');

Error using matlab.ui.control.internal.controller.ComponentController/executeUserCallback (line 378)
Error while evaluating DropDown PrivateValueChangedFcn.
```

I also received this error when attempting to run the `query_single` script from the command line.  Because most calculations are wrapped in try/catch blocks, I was unable to debug this further.

**Issue 3**

Insidious `clear all`, `close all`, and `clc` usage.  Many of the primary scripts and functions include these statements, including the setup script (`su.m`) and all the main query scripts (`query_single`, `query_batch`, `query_all`).  There is absolutely no reason for this and it can infuriate users like me when you dare to not just modify but completely wipe out their workspace.  In general, I will not use code that modifies my existing workspace without warning (and that includes touching my exisiting workspace variables, command line history, or existing figures).  I commented out the lines where I found them, but still found things being cleared and closed whenever I tried to test run a script, so I assume these calls are buried in many of the lower-level functions as well.  I cannot continue testing until these are all removed.  If you need a clear dedicated  workspace, move the calculations to a function.  If the calculations are already in a function, then why on earth would you need to clear things?!

**Issue 4**

No header documentation provided.  All functions should include help text following Matlab's [suggested format](https://www.mathworks.com/help/matlab/matlab_prog/add-help-for-your-program.html).  At minimum, I would expect an H1 line, syntax description, and a description of each input and output argument (including class, size, etc.).  It's very hard to test code when you have no idea what sort of inputs the functions are expecting.
