### 2021-02-18 #3002 @somefunAgba
@kakearney thanks for the robust feedback.
Issue 1 is the same with Issue 2. The problem being single quoted arguments do not allow double quoted calls in versions below R200a. I have fixed this.
Issue 3: Thanks for that, I will comment out hose housekeeping functions. They are only in a few files. Used them to keep my workspace less cluttered, kind of a working preference for me. I easily get irritated otherwise.
Also all calculations are done inside defined functions with a function file. To help with testing, I am adding automated tests as @agahkarakuzu recommended.
Issue 4:  I will add the minimum header to those functions without one.
