### 2021-02-20 475e66c
major refactorings added automated tests
