+++ cvmain/cvcore/upd_cdata_applet.m
@@ -1,22 +1,53 @@
-function upd_cdata_applet(app, update, search_ccode)
-%  UPD_CDATA Update local World Health Organization's COVID-19 dataset and process country 
-% by country data using country codes by specifying whether to update or not.
-% 
-% Uses last updated local dataset if not connected to the internet. This is 
-% a private function, used in the |get_cdata| public function.
-% 
-% |upd_cdata(1);| update and process all country data 
-% 
-% |upd_cdata(1, "NG");| update and process "NG" data
-% 
-% |upd_cdata(0, "US");| process but do not update "US" data
-% Copyright
+function status = upd_cdata_applet(search_ccode, update, app)
+%UPD_CDATA_APPLET Update COVID-19 WHO dataset 
+% Update and process country-code data.
+%
+% Uses last updated local dataset if not connected to the internet. 
+% This is a private function.
+%
+%INPUTS
+% (Required)
+% search_ccode : country code
+% (Optional)
+% update : update logic, 0 or 1
+% app : app handle
+%
+%OUTPUTS
+% status : success or failure, 1 or 0
+%
+%USAGE:
+% CMD
+% |upd_cdata_applet("ALL");| update and process all country-code data
+%
+% |upd_cdata_applet("NG",0);| process but do not update "NG" data
+%
+% |upd_cdata_applet("US");| update and process "US" data
+%
+% GUI
+% |upd_cdata_applet("ALL",1,app);| update and process all country-code data
+%
+% |upd_cdata_applet("NG",0,app);| process but do not update "NG" data
+%
+% |upd_cdata_applet("US",1,app);| update and process "US" data
+%
+%Copyright
 % <mailto:oasomefun@futa.edu.ng |oasomefun@futa.edu.ng|>|, 2020.|
 
-if nargin < 2
-    search_ccode = "ALL";
-end
+assert(nargin<=3," Expected at most 3 arguments!")
 
+is_app = true;
+if nargin < 3 % not an applet
+    is_app = false;
+    if nargin < 2
+        update = 1;
+    end
+end
+assert(update == 0 || update == 1, "update is either: 0 or 1!")
+
+ % default error state
+status = 0;
+
+% DB file names
 xlsx_file = "globalCV19_data.xlsx";
 cbc_CV19datafile = "cbc_CV19_data.xlsx";
 world_CV19datafile = "world_CV19_data.xlsx";
@@ -88,37 +119,50 @@ state = isfile(cbc_filel) && isfile(who_filel) && isfile(world_filel);
 
 %% import latest WHO COVID-19 data
 try
+    if is_app == true
+        app.StatusLabel.Text = "Querying ...";
+        app.StatusLabel.FontColor = [0.9, 0.5, 0.5];
+    end
+    
     if update == 1
         % Go online or fall back to last local copy.
         url="https://covid19.who.int/WHO-COVID-19-global-data.csv";
         
         try
-            websave("tmp.csv",url);
+            websave('tmp.csv',url);
             % convert csv file to table structure,
             % convert table structure to a excel's xlsx file
             % convert xlsx to table
-            T = readtable("tmp.csv","ReadVariableNames",true,...
-                "PreserveVariableNames",true,"TextType","string");
+            T = readtable('tmp.csv','ReadVariableNames',true,...
+                'PreserveVariableNames',true,'TextType','string');
             
             writetable(T,who_filed);
             % copy updated copy of "xlsx_file" in dir:data to dir:local
             copyfile(who_filed, who_filel, 'f');      
-            webaccess = true;            
+            webaccess = true;          
+            status = 1; % webaccess
             
-        catch
+        catch ME
             
             if state
                 webaccess = false;
-                skyblue = [0.5,0.7,0.9];
[…]
