### 2021-02-24 #3002 @kakearney
After pulling the most recent version of the code and reinstalling the app, I'm still getting the same errors when trying to use the GUI.  Both errors appear to be due to lack of backward compatibility.

The first error appears to originate from the `save_metrics` function, which specifies a `'WriteMode'` parameter when it calls `writetable`.  The `'WriteMode'` option was introduced in Matlab 2020a.

The second one originates from a call to the `sheetnames` function; `sheetnames` was introduced in 2019b.

I am currently running R2019a.  I do see that the README specifies that the code is only intended to be compatible back to 2019b, so the second error is my fault for overlooking that. I will try to update my version of Matlab in the next few days and retest.  In the meantime, I suggest adding a version check within the code itself so that it can throw an informative error if a user tries to run it on an older version of Matlab.

```matlab
if verLessThan('matlab', '9.7') % 9.7 = R2019b
    error('NLSIG-COVID19Lab requires Matlab R2019b or later');
end
```

As an additional note, the app appears to change the working directory when running, and when it throws an error, the user is stranded in a subfolder (in my example case, in the NLSIGCOVID19Lab/measures/infs/15-Dec-2020 folder).  I recommend looking into whether the code can be reworked to avoid changing the working directory.  If that cannot be avoided, it would be nice if the code took note of the original working directory and reset it when an error is caught.
