### 2021-02-24 #3002 @majensen
I will +1 the comments of @kakearney that in general caution against the application changing any aspect of the user's environment without notification. Apply the [Principle of Least Surprise](https://en.wikipedia.org/wiki/Principle_of_least_astonishment) and think in terms of a user who is completely ignorant and infinitely intelligent, and set that user's expectations explicitly through feedback.
