+++ cvmain/cvcore/upd_cdata_applet.m
@@ -1,8 +1,8 @@
 function status = upd_cdata_applet(search_ccode, update, app)
-%UPD_CDATA_APPLET Update COVID-19 WHO dataset 
+%UPD_CDATA_APPLET Update COVID-19 WHO dataset
 % Update and process country-code data.
 %
-% Uses last updated local dataset if not connected to the internet. 
+% Uses last updated local dataset if not connected to the internet.
 % This is a private function.
 %
 %INPUTS
@@ -33,6 +33,7 @@ function status = upd_cdata_applet(search_ccode, update, app)
 %Copyright
 % <mailto:oasomefun@futa.edu.ng |oasomefun@futa.edu.ng|>|, 2020.|
 
+
 assert(nargin<=3," Expected at most 3 arguments!")
 
 is_app = true;
@@ -44,7 +45,7 @@ if nargin < 3 % not an applet
 end
 assert(update == 0 || update == 1, "update is either: 0 or 1!")
 
- % default error state
+% default error state
 status = 0;
 
 % DB file names
@@ -58,7 +59,7 @@ if ~(ismcc || isdeployed)
     [thisfp,thisfn,~]= fileparts(which('upd_cdata_applet.m'));
     rootfp = strrep(thisfp, [filesep 'cvmain' filesep 'cvcore'], '');
     if isfile(fullfile(thisfp,thisfn+".m"))
-        old_dir = cd(rootfp);
+        current_userfp = cd(rootfp);
     end
 else
     % we don't need to do anything
@@ -71,10 +72,10 @@ if ~(ismcc || isdeployed)
     if update == 0 && search_ccode == "ALL"
         e_msg = sprintf("Nothing to do.\n");
         cprintf('[0.5, 0.5, 0.5]',char(e_msg));
-        cd(old_dir);
+        cd(current_userfp);
         return
     end
-        
+    
     who_filel = fullfile(rootfp, "local", xlsx_file);
     cbc_filel = fullfile(rootfp, "local", cbc_CV19datafile);
     world_filel = fullfile(rootfp, "local", world_CV19datafile);
@@ -83,14 +84,15 @@ if ~(ismcc || isdeployed)
     who_filed = fullfile(rootfp, "data", xlsx_file);
     cbc_filed = fullfile(rootfp, "data", cbc_CV19datafile);
     world_filed = fullfile(rootfp, "data", world_CV19datafile);
-
-    % create data dir.
-    try
-        old_folder = cd('data');
-    catch
-        mkstore("data");
-        old_folder = cd("data");
-    end
+    
+    %     % create data dir.
+    %     try
+    %         old_folder = cd('data');
+    %     catch
+    %         mkstore("data");
+    %         old_folder = cd("data");
+    %     end
+    cd(current_userfp);
 else
     
     if update == 0 && search_ccode == "ALL"
@@ -98,7 +100,7 @@ else
         app.StatusLabel.Text = e_msg;
         app.StatusLabel.FontColor = [0.5, 0.5, 0.5];
         if ~(ismcc || isdeployed)
-            cd(old_dir);
+            cd(current_userfp);
         end
         return
     end
@@ -117,10 +119,11 @@ end
 % check if the 'xlsx_file' and'cbc_CV19datafile' exists in local folder
 state = isfile(cbc_filel) && isfile(who_filel) && isfile(world_filel);
 
+pause('on')
 %% import latest WHO COVID-19 data
 try
     if is_app == true
-        app.StatusLabel.Text = "Querying ...";
+        app.StatusLabel.Text = "Querying web ...";
         app.StatusLabel.FontColor = [0.9, 0.5, 0.5];
     end
     
@@ -138,8 +141,8 @@ try
             
             writetable(T,who_filed);
[…]
