+++ cvmain/ipfinder/xy_ipsort.m
@@ -29,7 +29,10 @@ for id = 1:n_ips
     % debug: disp(id)
     % CASE 1.
     xmins(id) = x(ips_valley(id));
-    ymins(id) = y(ips_valley(id));
+    ymins(id) = y(ips_valley(id));        
+    if id == 1
+            ymins(id) = y(1);
+    end
     try
         xpks(id) = x(ips_peak(id));
         xmaxs(id) = x(ips_valley(id+1));
@@ -58,11 +61,15 @@ for id = 1:n_ips
                %  xmaxs(id) = xpks(id)*ymaxs(id)/(ypk);
                 xmaxs(id) =  xmins(id) + ...
                     ((xpks(id)-xmins(id))*(ymaxs(id)-ymins(id))/(ypk-ymins(id)));
+                if id == n_ips
+                    ymaxs(id) = y(end);
+                end
             catch
                 error("Bad...something wrong occured!")
             end
         end
     end
+
 end
 
 

[…]
