### 2021-03-01 a3ec740
minor refactorings
