### 2021-03-01 106da80
Set up CI with Azure Pipelines
[skip ci]
