### 2021-03-01 1d12c86
Update azure-pipeline.yml for Azure Pipelines
