### 2021-03-01 a6a866c
Update azure-pipeline.yml for Azure Pipelines
