### 2021-03-01 cc84e11
Update azure-pipeline.yml for Azure Pipelines
