### 2021-03-01 419877d
Update azure-pipeline.yml for Azure Pipelines
