+++ cvmain/nlsigcore/lnn.m
@@ -1,1382 +1,0 @@
-classdef lnn < handle
-    %LNN Logistic Neural Network
-    %   General SISO/MISO to SIMO/MIMO case of the NLSIG neural pipeline
-    
-    properties
-        % data: explanatory input patterns: L x D
-        Xin;
-        % data: actual outcomes: J x D
-        RYout;
-        % lnn: inferred outcomes: J x D
-        Yout;
-        % lnn: input layer weights/critics: J x (L+1)
-        Win;
-        
-        % lnn: tune weight bias, w0 can be false (default) or true
-        tune_wbias = 0;
-        % tune input weight can be false (0) or true (1)
-        tune_win = 0;
-        
-        % lnn: mode for supervised learning, "s" | 1
-        % or unsupervised learning, "us" | 0
-        lrnmode;
-        
-        % lnn: mode for regression or classification-based regression
-        % : "genreg" | 1 , "classreg" | 0
-        fitmode;
-        
-        %
-        L; % number of explanatory inputs
-        J; % number of categorical outcomes
-        D; % length of data
-        N; % array of ns.
-        
-        %
-        batch_size; % default is 1
-        type_train; % training type - stochastic (online), batch or mixed
-        optimtype; % optimization type (typically gradient-descent)
-        isshuffle = false;
-        
-        % nlsig
-        np; % max. order of output derivatives
-        eq; % equal or general intervals
-        % nlsig neuron options
-        nlsigopts;
-        p; % number of estimated parameters in each ith partition
-        % input est.parameters structure
-        solp;
-        % fitted solp
-        solpO;
-        % dkw lower-bnd solp
-        solpL;
-        % dkw upper-bnd solp
-        solpH;
-        % vectorized est.parameters structure
-        % version 1
-        % assumes 1:nj for each p=1:P in each j
-        % version 2
-        % assumes 1:pi for each i=1:nj in each j
-        solpvec;
-        % unroll version
-        urollvs;
-        
-        % state placeholder
-        x; % input data
-        u; % temporal input data
-        ry; % correct output data
-        
-        % D x J
-        E; % error output at logistic layer
-        Ess; % sum of square error objective at logistic layer
-        DYDx; % first derivative of the output of each logistic layer
-        DYDxnp;% 1:np-th derivative of the output of each logistic layer
-        % JH: jacobian-hessian ds
-        JH;
-        
-        % dj x pnj
-        JACLY1; JACLE1; JACLEss1;
-        % pnj x pnj
-        HESLY1; HESLE1; HESLEss1;
-        % j x pnj
-        JACLY2; JACLE2; JACLEss2;
-        % pnj x pnj
-        HESLY2; HESLE2; HESLEss2;
-        % 1 x pnj
-        JACLY3; JACLE3; JACLEss3;
-        % pnj x pnj
-        HESLY3; HESLE3; HESLEss3;
-        %
-        Fout
-        Jout;
-        Hout;
-        
-        % D X (L+1) X J
-        pDYDW1;pDEDW1;pDEssDW1;
-        % J * (L+1)
-        pDYDW2; pDEDW2;pDEssDW2;
-        % (L+1) x (L+1)
-        HESWY;HESWE,HESWEss
-        % (L+1) x (J)
-        GRADWEss;
-        
-        % D X J
-        pDEssDY;
-        pDEssDx;
-        
-        % dj x lj
-        JACWY1; JACWE1; JACWEss1;
-        % lj x lj
-        HESWY1; HESWE1; HESWEss1;
-        % j x lj
-        JACWY2; JACWE2; JACWEss2;
-        % lj x lj
-        HESWY2; HESWE2; HESWEss2;
-        % 1 x lj
[…]
