+++ cvmain/nlsigcore/tests/inflx_testi.m
@@ -1,10 +1,0 @@
-lnn_testi;
-infl = inflxpt;
-tpasses = 1;
-infl.setprops(tpasses,lnnet.Yout,x);
-infl.find();
-[infl,phase]=infl.group();
-p = 7;
-eq = 0;
-base = "nat";
-[infl,solpi]=infl.sort(false,p,eq,base);

[…]
