### 2021-03-01 d0c129e
Set up CI with Azure Pipelines
[skip ci]
