### 2021-03-01 cdad3d0
Update azure-pipeline.yml for Azure Pipelines
