### 2021-03-01 6f003b0
Update azure-pipeline.yml for Azure Pipelines
