### 2021-03-01 73b3fa2
Update azure-pipeline.yml for Azure Pipelines
