### 2021-03-01 10226aa
Update azure-pipeline.yml for Azure Pipelines
