+++ azure-pipeline.yml
@@ -5,7 +5,7 @@ pool:
 steps:
   - task: InstallMATLAB@0
     inputs:
-      release: 'R2020b'
+      release: 
   - task: RunMATLABTests@0
     inputs:
       #sourceFolder: root;
@@ -25,4 +25,3 @@ steps:
     inputs:
       codeCoverageTool: Cobertura
       summaryFileLocation: code-coverage/coverage.xml
-      
