+++ cvmain/cvcore/get_cdata_applet.m
@@ -27,8 +27,8 @@ function [TT,status] = get_cdata_applet(search_ccode, update, app)
 %Copyright:
 % <mailto:oasomefun@futa.edu.ng |oasomefun@futa.edu.ng|>|, 2020.|
 
-if verLessThan('matlab', '9.7') % 9.7 = R2019b
-    error('NLSIG-COVID19Lab requires Matlab R2019b or later');
+if verLessThan('matlab', '9.8') % 9.7 = R2019b
+    error('NLSIG-COVID19Lab requires Matlab R2020a or later');
 end
 
 assert(nargin<=3," Expected at most 3 arguments!")
