### 2021-03-02 8f62f67
Update testAppUpdnModel.m
