### 2021-03-12 #3002 @kakearney
I updated my version of Matlab to R2020b and downloaded the most recent version of the NLSIG-COVID19Lab tool, and was now able to complete my review.

To test it, I first played around with the App; I ran simulations using the default parameters and also tested entering new values in each of the dropdown menus.  This new version of the app is well-designed, with a straightforward user interface and restrictions of the input that prevent the user from entering nonsense values as input... well done!

I also ran through the examples in the example folder, using the live script version, and ran the test suite.  Everything appeared to run smoothly, although it did take nearly 3 hours to complete the tests.  I did hit one error, described below, but I believe that is due to a typo in the test script rather than errant behavior of the code itself.

Overall, the author appears to have addressed my comments regarding respecting the user's workspace.  The use of relative path names has replaced many directory changes, and with the exception of the setup script, calls to `clear` and `clc` have been removed.  Thank you.

I was unable to check off items in the menu above, but have added my checklist at the bottom of this comment, with a few clarifications in italics.  I believe the remaining issues are minor, and should not prevent publication of this entry.

## Remaining issues

### Warnings when exiting app

When closing the App, I received the following dump of warning messages.  They appear harmless, but it would be nice to eliminate them if possible.

```matlab
Warning: "/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/assets" not
found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning:
"/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/assets/01-Sep-2020"
not found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning:
"/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/assets/21-Feb-2021"
not found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning:
"/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/assets/22-Feb-2021"
not found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning: "/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/measures"
not found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning:
"/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/measures/dths" not
found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning:
"/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/measures/dths/22-Feb-2021"
not found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning:
"/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/measures/infs" not
found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning:
"/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/measures/infs/01-Sep-2020"
not found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning:
"/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/measures/infs/21-Feb-2021"
not found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
Warning:
"/Users/kakearney/Documents/MATLAB/Add-Ons/Apps/NLSIGCOVID19Lab/measures/infs/22-Feb-2021"
not found in path.
> In rmpath (line 71)
In appinstall.internal.stopapp>cleanup (line 38)
In appinstall.internal.stopapp>cleanup13a (line 27)
In appinstall.internal.stopapp (line 4)
In NLSIGCOVID19LabApp>@()appinstall.internal.stopapp([],[],obj) (line 108)
In onCleanup/delete (line 80)
In closereq (line 18)
In matlab.ui.internal.controller.FigureController>@(o,e)this.Model.hgclose() (line 381)
In internal/Callback/execute (line 128)
In matlab.internal.cef/webwindow/onCustomEvent (line 1348)
In matlab.internal.cef.webwindow>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 351)
In asyncio/Channel/onCustomEvent (line 477)
In asyncio.Channel>@(source,data)obj.onCustomEvent(data.Type,data.Data) (line 409)
>>
```

### Test suite

I did receive one failure in the testAppUpdnModel test:

```
================================================================================
Verification failed in testAppUpdnModel/test_ModelButton.
    ---------------------
    Framework Diagnostic:
    ---------------------
    verifyEqual failed.
    --> The objects are not equal using "isequaln".

    Actual Value:
      datetime

       01-Sep-2020
    Expected Value:
      datetime

       01-May-2020
    ------------------
    Stack Information:
    ------------------
    In <mypath_redacted>/NLSIG-COVID19Lab/tests/testAppUpdnModel.m (testAppUpdnModel.test_ModelButton) at 65
================================================================================
```

The failure is the result of a possible typo in testAppUpdnModel.m, lines 64-65; the set and verify values are different, which causes that verification test to fail.

### Paper

I had to look at some references to get a better understanding of what the XIR and YIR metrics meant (i.e. that they quantified inflection points in the curve).  I suggest adding a line or two of text to the "NLSIG Preprint" link at the top of the readme to let readers know that that's where they should go to find the in-depth discussion of the math underlying this tool and how to best interpret results.

Also, currently, the link to the paper in the README is broken; I assume that will be fixed when the paper is finalized.

## Checklist

*I have placed an X at the start of each checked-off bullet, and put some comments in italics*

**Conflict of interest**

- X I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

**Code of Conduct**

- X I confirm that I read and will adhere to the JOSS code of conduct.

**General checks**

- X Repository: Is the source code for this software available at the repository url?
- X License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- X Contribution and authorship: Has the submitting author (@somefunAgba) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
- X Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

**Functionality**

- X Installation: Does installation proceed as outlined in the documentation?
- X Functionality: Have the functional claims of the software been confirmed?
- X Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

**Documentation**

- X A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

  *The target audience as stated to be both researchers and non-researchers.  Given the niche techniques and terminology, and the reliance on a technical computing language like Matlab, this is most likely to appeal to a pretty specialized audience.  However, as an outsider to the field, I was able to at least play around with the tool a bit, and given the widespread interest in looking at COVID data, this the wide target audience is not unreasonable.*
- X Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
- X Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
- X Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
- X Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

  *I do not see any explicit instructions for collaboration in the README or paper, though the author has been responsive to issues submitted through the GitHub interface.*

**Software paper**

- X Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
- X A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
- X State of the field: Do the authors describe how this software compares to other commonly-used packages?
- X Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

  *The are some minor grammatical issues in the paper (e.g. sentence fragments) but these did not prevent understanding of the paper.  Some copy-editing may be necessary before final publication.*
- X References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

  *I will reiterate that this topic is beyond my area of expertise, so I am not certain whether there are additional references related to the topic.  However, existing citations appear comprehensive, and I was able to easily track down answers to some questions I had while playing around with the software (specifically, looking for more info on the XIR and YIR metrics; more context about methods used for predicting COVID curves) based on the reference list in the paper.*
