### 2021-03-13 f607899
cleanup
fix: warnings on closing app
