### 2021-03-13 9eefb15
Update exportplots_applet.m
