+++ cvmain/plotter/plotpreds_applet.m
@@ -1,6 +1,6 @@
 function plotpreds_applet(app,country_code,time_data,x_data,y_data,dy_data,...
     y_sol,y_sollb,y_solub,dy_sol,dy_sollb,dy_solub,...
-    focus,ips_adata,ips_vdata,ips_pdata,phase)
+    focus,ips_adata,ips_vdata,ips_pdata,phase) %#ok<INUSL>
 %PLOTPREDS nlsig Predictions Plots
 
 try
@@ -35,8 +35,8 @@ plot(app.UIAxes,time_data(ips_pdata),y_data(ips_pdata),'+')
 
 
 plot(app.UIAxes,time_data,y_data,'MarkerSize',3,'Color',lightgrey1,'LineWidth',0.5);
-plot(app.UIAxes,time_data,y_sollb,'-.','Color',khaki4,'LineWidth',1);
-plot(app.UIAxes,time_data,y_solub,'-.','Color',khaki3,'LineWidth',1);
+% plot(app.UIAxes,time_data,y_sollb,'-.','Color',khaki4,'LineWidth',1);
+% plot(app.UIAxes,time_data,y_solub,'-.','Color',khaki3,'LineWidth',1);
 plot(app.UIAxes,time_data,y_sol,'Color',red2,'LineWidth',1.2);
 
 
@@ -65,8 +65,8 @@ bar(app.UIAxes2, dy_data,0.5,'EdgeColor','none',...
     'FaceColor',0.65*ones(1,3),'FaceAlpha',0.1);
 
 plot(app.UIAxes2,iplines,dy_data,'LineWidth',0.15,'Color',[0.9 0.9 0.9]);
-plot(app.UIAxes2,x_data, dy_sollb,'-.','Color',khaki4,'LineWidth',1);
-plot(app.UIAxes2,x_data, dy_solub,'-.','Color',khaki3,'LineWidth',1);
+% plot(app.UIAxes2,x_data, dy_sollb,'-.','Color',khaki4,'LineWidth',1);
+% plot(app.UIAxes2,x_data, dy_solub,'-.','Color',khaki3,'LineWidth',1);
 plot(app.UIAxes2,x_data, dy_sol,'Color',red2,'LineWidth',1.2);
 
 
