+++ cvmain/cvcore/cov19_nlsigquery_applet.m
@@ -215,6 +215,7 @@ end
 pause('on');
 %% Second-Pass 
 % Uses smoothed solution data to ensures metrics are correct
+smooth = false;
 while (smooth == true)
     spass = true;
     smooth  = false;
@@ -280,6 +281,7 @@ while (smooth == true)
     
 end
 pause('on');
+
 %% NLSIG curve metrics
 if spass == false
     YIR = YIRidx(gy_sol,gsol.ymax,gsol.ymin);
