### 2021-03-22 #3002 @agahkarakuzu
@somefunAgba I had to upgrade my MATLAB version to R2020b to meet the minimum version requirement, sorry for the delay.

I was easily able to follow the new instructions in the new README and produce expected outputs. I think users with little matlab experience would be able to do so. Thank you so much for improving the documentation and user experience.

I also see that you set up [Azure pipelines](https://dev.azure.com/oasomefun/NLSIGCOVID19_LAB/_build?definitionId=3) for the CI tests, but they are not passing yet due to double quotes issue, for example in `testScriptUpdData.m`:

```octave
status = upd_status("WD");
```

> The problem being single quoted arguments do not allow double quoted calls in versions below R2020a. I have fixed this.

I see that your config file requests `R2020b` from Azure, but looks like this is still an issue. Despite that Azure installs the latest unless specified, I am suspicious that the `R2020b` installation did not take effect. Somewhere in your test can you ensure that the requested version matches the installed one by calling the `version` and maybe try:

```yaml
steps:
  - task: InstallMATLAB@0
    inputs:
      release: R2020b
```

Another option would be simply changing double quotes with single quotes in the test script and see if the tests pass on Azure. I think it is important to have successful CI builds as well as having them set up. This is the last item on my checklist to cross off.

I am pretty much in the same boat with @kakearney regarding the scientific content of the article. Although some of the content went above my head, I was still able to understand where does this software stand in relation to the existing ones.
