+++ tests/testAppUpdnModel.m
@@ -14,7 +14,7 @@ classdef testAppUpdnModel < matlab.uitest.TestCase
             [this_filepath,this_filename,~]= ...
                 fileparts(mfilename('fullpath')); %#ok<ASGLU>
             %rootpath = this_filepath;
-            rootpath = strrep(this_filepath, [filesep 'utests'], '');
+            rootpath = strrep(this_filepath, [filesep 'tests'], '');
             addpath(genpath(rootpath));
             if isfolder(fullfile(rootpath,'bin'))
                 rmpath(fullfile(rootpath,"bin"))
