+++ tests/testScriptQueryModel.m
@@ -11,7 +11,7 @@ classdef testScriptQueryModel < matlab.unittest.TestCase
             [this_filepath,this_filename,~]= ...
                 fileparts(mfilename('fullpath')); %#ok<ASGLU>
             %rootpath = this_filepath;
-            rootpath = strrep(this_filepath, [filesep 'utests'], '');
+            rootpath = strrep(this_filepath, [filesep 'tests'], '');
             addpath(genpath(rootpath));
             if isfolder(fullfile(rootpath,'bin'))
                 rmpath(fullfile(rootpath,"bin"))
