### 2021-03-23 #3002 @somefunAgba
Thanks @agahkarakuzu.
I have set the release in the azure pipeline file to R2020b.

I just discovered that the reason, why the tests were failing. I also was wondering if it was not a typo issue or release issue, till I gave up. This morning I realized that it was due to a **path typo** in each of the unit-test files in the test-suite folder. I have fixed this now.

Currently, the only test-suite that did not pass was `testAppUpdnModel`. This test-suite invokes the app, which requires that a display be present in order to test the function properly. **Since this cannot be currently achieved, it fails fast**.  The problem is described and explained in the link below:
[https://www.mathworks.com/matlabcentral/answers/441139-why-am-i-getting-an-error-regarding-matlab-uiautomation-driver-nodisplay-when-i-run-matlab-unit-test](https://www.mathworks.com/matlabcentral/answers/441139-why-am-i-getting-an-error-regarding-matlab-uiautomation-driver-nodisplay-when-i-run-matlab-unit-test)

In this case:

Error occurred in testAppUpdnModel/test_UpdButton and it did not run to completion.
    ---------
    Error ID:
    ---------
    'MATLAB:uiautomation:Driver:NoDisplay'
    --------------
    Error Details:
    --------------
    Error using matlab.ui.internal.Driver/press (line 8)
    Gestures on MATLAB Apps require figure display.

    Error in matlab.uitest.internal.GestureProvider/press (line 44)
                testCase.Driver.press(H, varargin{:});

    Error in testAppUpdnModel/test_UpdButton (line 44)
                testCase.press(testCase.App.UpdateDatabaseButton)

**The other tests that do not require a graphics display functionality are passing**.
