### 2021-03-23 4e894b9
update app test functionality to reflect no-display mode
