+++ tests/units/testScriptUpdData.m
@@ -11,7 +11,7 @@ classdef testScriptUpdData < matlab.unittest.TestCase
             [this_filepath,this_filename,~]= ...
                 fileparts(mfilename('fullpath')); %#ok<ASGLU>
             %rootpath = this_filepath;
-            rootpath = strrep(this_filepath, [filesep 'tests'], '');
+            rootpath = strrep(this_filepath, [filesep 'tests' filesep 'units'], '');
             addpath(genpath(rootpath));
             if isfolder(fullfile(rootpath,'bin'))
                 rmpath(fullfile(rootpath,"bin"))
@@ -23,7 +23,7 @@ classdef testScriptUpdData < matlab.unittest.TestCase
         end
     end
     
-    methods(Test)
+    methods(Test, TestTags = {'Unit'})
         % unit-test functions
         
         function testUpdDat1(testCase)
@@ -36,7 +36,10 @@ classdef testScriptUpdData < matlab.unittest.TestCase
             % check that: no exception occurs
             testCase.verifyNotEqual(status,0);            
         end
-        
+    end
+    
+    methods(Test)
+    
         function testUpdDat2(testCase)
             %TESTUPDDAT2
             %   Re-Test if local-database can be
