+++ azure-pipeline.yml
@@ -12,7 +12,8 @@ steps:
     inputs:
       #sourceFolder: root;
       #sourceFolder: root/folderA; root/folderB
-      selectByFolder: 'tests'
+      selectByFolder: 'tests/units'
+      selectByTag: Unit
       testResultsPDF: test-results/results.pdf
       testResultsJUnit: test-results/results.xml
       codeCoverageCobertura: code-coverage/coverage.xml
