### 2021-03-23 #3002 @somefunAgba
Ok. Quick update @agahkarakuzu.

**All the CI tests on Azure are now passing.**

I have fixed the gaphics display error by adding a test
to assert whether matlab is being run in a nodisplay mode or display mode
