### 2021-03-23 #3002 @agahkarakuzu
@somefunAgba that's great! Yes, I am aware of that display limitation, but happy that all the other issues are fixed easily!

This was the last item on my end, you may want to add badges for coverage and CI status to README if you see fit.

With these changes, I checked the last box.
