### 2021-03-23 #3002 @majensen
@kakearney @agahkarakuzu really appreciate your careful technical reviews and @somefunAgba all the work to fix issues. I know both reviewers indicated less familiarity with the underlying mathematics; I will weigh in on this for the record (spoiler alert: looks ok and also very interesting) for the record. I will also have a readthrough of the paper itself and may make a pull request with editorial suggestions.
Hope I can complete these tasks over the next week.
