### 2021-04-03 #3002 @majensen
@somefunAgba - there are number of references that Whedon found DOIs for (see [above](https://github.com/openjournals/joss-reviews/issues/3002#issuecomment-812871224)). Could you please add these to your Ref section and bib? Thanks a lot-
