### 2021-04-03 #3002 @majensen
@somefunAgba, here is my understanding and overall impression of the method itself. Welcome your comments.

My impression of the story.

Logistic growth is characteristic of many, many natural phenomena. It captures key real world system behaviors of a maximal growth rate and also of diminishing returns. As a growth system's external influences change, logistic growth should remain a good representation of the new system behavior, provided that the growth parameters are adjusted to the changing influences. In fact, one can imagine adding together (superposing) multiple logistic functions of appropriate parameters, appropriately translated along the domain axis, to estimate or fit any time series whose underlying process is one of growth constrained by limited resources.

Fitting an appropriate superposition of logistic functions to data is _based on a model of the underlying process_, in contrast to other forms of estimation (such as stitching together splines at their endpoints), whose goal is to minimize the difference between data points and an arbitrary smooth function. The authors make a key insight, realizing that superposition of identically-parametered logistic sigmoid functions is equivalent to a certain neural network, the activation function of whose nodes are precisely the functions to be fitted. This, and the fact that the functions are of identical form, enables the authors to bring to bear numerical neural network theory on the paramter fitting problem - which becomes an optimization problem for which, in these days of deep learning, has many efficient tools.

Another key insight of the method is the reduction of dimensionality, by fitting, not _all_ data points, but only points of sigmoid inflection. This takes advantage of the assumption of the underlying process (logistic growth) driving the dynamics, and presumably leads to a better foundation for a mechanistic interpretation of the fit.

The authors are able to make use of a variant of the Kolmogorov-Smirnov inequality to derive meaningful assessments of goodness of fit and confidence intervals on the point estimations of the time series. In my opinion, the goodness-of-fit results, at least in the examples provided in the paper, are a vindication of the method. But the logic of [the derivation](https://arxiv.org/abs/2008.04210) is also clear and seems sound to me.
