+++ osspaper/paper.md
@@ -33,7 +33,7 @@ model in Facebook's Prophet model for time-series growth-forecasting
 at scale [@taylorForecastingScale2018] on big data. 
 The scientific basis for this prevalence is given in [@bejanConstructalLawOrigin2011]. 
 Such growth-processes can be viewed as complex input--output systems that involve 
-multiple peak inflection phases with respect to time. An idea that 
+multiple peak inflection phases with respect to time, an idea that 
 can be traced back in the crudest sense to [@reedSummationLogisticCurves1927]. A modern definition for the logistic-sigmoid growth which considers restricted growth from  a two-dimensional perspective is the nlogistic-sigmoid function (`NLSIG`) [@somefunLogisticsigmoidNlogisticsigmoidModelling2020] 
 or logistic neural-network (`LNN`) pipeline. 
 
@@ -43,14 +43,14 @@ of the COVID-19 epidemic growth in each affected country of the world and the wo
 
 # Statement of need
 
-Admittedly, epidemiological models such as the SEIRD variants 
-[@leeEstimationCOVID19Spread2020;@okabeMathematicalModelEpidemics2020] are just another form of representing sigmoidal growth [@xsRichardsModelRevisited2012]. It has been noted in 
+Epidemiological models such as the SEIRD variants 
+[@leeEstimationCOVID19Spread2020;@okabeMathematicalModelEpidemics2020] are just another form of representing sigmoidal growth [@xsRichardsModelRevisited2012]. However, it has been noted in 
 [@christopoulosNovelApproachEstimating2020] that the SEIRD-variant models yield largely exaggerated forecasts. 
-Observing the current state of the COVID-19 pandemic, this is also true, as regards, 
+Observing the current state of the COVID-19 pandemic, this is concern is borne out, in 
 the results of various applications of logistic modelling [@batistaEstimationStateCorona2020;@wuGeneralizedLogisticGrowth2020]
-which have largely resulted in erroneous identification of the epidemic's progress and its future projection, hence leading policymakers astray [@matthewWhyModelingSpread2020]. 
+which have largely resulted in erroneous assessment of the epidemic's progress and its future projection, hence leading policymakers astray [@matthewWhyModelingSpread2020]. 
 
-Notably, two recurring limitations of the logistic definitions in the literature and other software packages exist. These two limitations are trends that have persisted since the first logistic-sigmoid function introduction [@bacaerVerhulstLogisticEquation2011]. 
+Notably, two recurring limitations of the logistic definitions in the literature and other software packages exist. These two limitations are trends that have persisted since the first introduction of the logistic-sigmoid function [@bacaerVerhulstLogisticEquation2011]. 
 
 First is that, the co-domain of logistic function is assumed to be infinite. This assumption violates the natural principle of finite growth. 
 Second is that, during optimization, estimation of the logistic hyper-parameters  for the individual logistic-sigmoids that make the multiple logistic-sigmoid sum is computed separately, instead of as a unified function. The effect of this, is that, as the number of logistic-sigmoids 
@@ -58,7 +58,7 @@ considered in the sum increases, regression analysis becomes more cumbersome and
 @hsiehRealtimeForecastMultiphase2006;@wuGeneralizedLogisticGrowth2020;
 @chowellNovelSubepidemicModeling2019;@taylorForecastingScale2018]. 
 
-These limitations are efficiently overcome by the nlogistic-sigmoid function `NLSIG` (or logistic neural-network pipeline) for describing logistic growth. We note that the `NLSIG` is a logistic neural-network machine-learning tool under active development. The benefits it provides at a functional level:
[…]
