+++ osspaper/paper.bib
@@ -1185,7 +1185,7 @@
 }
 
 @article{okabeMathematicalModelEpidemics2020,
-  title = {A {{Mathematical Model}} of {{Epidemics}}\textemdash{{A Tutorial}} for {{Students}}},
+  title = {A {{Mathematical Model}} of {{Epidemics}}---{{A Tutorial}} for {{Students}}},
   author = {Okabe, Yutaka and Shudo, Akira},
   year = {2020},
   month = jul,
