+++ osspaper/paper.md
@@ -44,34 +44,34 @@ of the COVID-19 epidemic growth in each affected country of the world and the wo
 # Statement of need
 
 Epidemiological models such as the SEIRD variants 
-[@leeEstimationCOVID19Spread2020;@okabeMathematicalModelEpidemics2020] are just another form of representing sigmoidal growth [@xsRichardsModelRevisited2012]. However, it has been noted in 
+[@leeEstimationCOVID19Spread2020;@okabeMathematicalModelEpidemics2020] are just another form of representing sigmoidal growth [@xsRichardsModelRevisited2012]. However, it has been noted 
 [@christopoulosNovelApproachEstimating2020] that the SEIRD-variant models yield largely exaggerated forecasts. 
 Observing the current state of the COVID-19 pandemic, this is concern is borne out, in 
 the results of various applications of logistic modelling [@batistaEstimationStateCorona2020;@wuGeneralizedLogisticGrowth2020]
-which have largely resulted in erroneous assessment of the epidemic's progress and its future projection, hence leading policymakers astray [@matthewWhyModelingSpread2020]. 
+which have largely led to erroneous assessments of the epidemic's progress and its future projection, leading policymakers astray [@matthewWhyModelingSpread2020]. 
 
 Notably, two recurring limitations of the logistic definitions in the literature and other software packages exist. These two limitations are trends that have persisted since the first introduction of the logistic-sigmoid function [@bacaerVerhulstLogisticEquation2011]. 
 
 First is that, the co-domain of logistic function is assumed to be infinite. This assumption violates the natural principle of finite growth. 
 Second is that, during optimization, estimation of the logistic hyper-parameters  for the individual logistic-sigmoids that make the multiple logistic-sigmoid sum is computed separately, instead of as a unified function. The effect of this, is that, as the number of logistic-sigmoids 
-considered in the sum increases, regression analysis becomes more cumbersome and complicated as can be observed in these works [@leeEstimationCOVID19Spread2020;@batistaEstimationStateCorona2020;
+considered in the sum increases, regression analysis becomes more cumbersome and complicated as can be observed in a number of works [@leeEstimationCOVID19Spread2020;@batistaEstimationStateCorona2020;
 @hsiehRealtimeForecastMultiphase2006;@wuGeneralizedLogisticGrowth2020;
 @chowellNovelSubepidemicModeling2019;@taylorForecastingScale2018]. 
 
 These limitations are efficiently overcome by the nlogistic-sigmoid function `NLSIG` (or logistic neural-network pipeline) for describing logistic growth. We note that the `NLSIG` is a logistic neural-network machine-learning tool under active development. The benefits it provides at a functional level are:
 	
- - unified function definition
+ - unified function definition,
  
- - functional simplicity and efficient computation
+ - functional simplicity and efficient computation,
 	
- - improved nonlinear modelling power
+ - improved nonlinear modelling power.
 		
 Ultimately, the development of the `NLSIG-COVID19Lab` was motivated by research needs, in that it 
-illustrates the power of the nlogistic-sigmoid neural pipeline. `NLSIG-COVID19Lab` provides an optimization workflow with functions to make modelling and monitoring the COVID-19 pandemic easier and reliable. Notably, instead of engaging in false prophecy 
+illustrates the power of the nlogistic-sigmoid neural pipeline. \linebreak `NLSIG-COVID19Lab` provides an optimization workflow with functions to make modelling and monitoring the COVID-19 pandemic easier and reliable. Notably, instead of engaging in false prophecy 
 or predictions on the cumulative growth of an ongoing growth phenomena, whose source is both uncertain and 
[…]
