### 2021-04-03 #3002 @majensen
I have made https://github.com/somefunAgba/NLSIG-COVID19Lab/pull/4 with some very minor edits for consideration.
