### 2021-04-04 #3002 @majensen
Ok @somefunAgba - the last thing to do before we request publication is to create an open archive of the software. This can be done using [Zenodo](https://zenodo.org), [FigShare](https://figshare.org), or similar service. There is a short tutorial at [this reference](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729).
Once you create the archive, can you please report the version number and the archive DOI back in this thread? Thank you.
