### 2021-04-05 #3002 @agahkarakuzu
@majensen that was such a helpful comment! Despite that logistic growth is not my bread and butter, your comment gave me a great starting point to see what is really cool about considering multiple inflection points sampled at irregular finite intervals when modeling the limits of the exponential growth. I gave [Somefun et al. 2020](https://arxiv.org/abs/2008.04210) another read with the encouragement from your high-level insight, then it all clicked into place, including the significance of using cumulative growth.

Looking forward to the time point when we are at the phase of perceiving covid-19 as a past finite incident, and remember how NLSIG provided an efficient solution to capture the true nature of its bounds, even while it was still unbearably continuous.

Thank you @somefunAgba for your contribution and careful consideration of the issues we raised about the codebase quality/ux with @kakearney, @majensen for the insightful summary of the methodology and JOSS for making all these conversations publicly available. This review proved me once again that reinforcing understanding by reading comments is a unique learning experience.
