### 2021-04-06 #3002 @majensen
@openjournals/joss-eics - Note that all references actually appearing the paper have working DOIs, except for

`Christopoulos, D. T. (2016). On the efficient identification of an inflection point. International
Journal of Mathematics and Scientific Computing,(ISSN: 2231-5330), 6(1).`

which does not have a DOI but is available online from the journal at https://www.veltech.edu.in/wp-content/uploads/2016/04/Paper-04-2016.pdf
