### 2021-04-06 #3002 @danielskatz
no problem @majensen - all references don't have to have DOIs in the bib, just the ones that actually have them irl...
