### 2021-04-06 #3002 @danielskatz
👋 @somefunAgba - one of your references, to "" says it is actually the same as 10.1109/MSPEC.2020.9205546 - which is how it was originally published. Can we change the reference to this original version (and add the DOI)?

I'll also suggest some more changes via a PR in a little bit
