+++ osspaper/paper.bib
@@ -1047,6 +1047,7 @@
   month = sep,
   url = {https://spectrum.ieee.org/artificial-intelligence/medical-ai/why-modeling-the-spread-of-covid19-is-so-damn-hard},
   urldate = {2020-12-04},
+  doi = {10.1109/MSPEC.2020.9205546},
   file = {Snapshot:C\:\\Users\\SomefunAgba\\Zotero\\storage\\PU7622RS\\why-modeling-the-spread-of-covid19-is-so-damn-hard.html:text/html},
   journal = {IEEE Spectrum: Technology, Engineering, and Science News},
   language = {en},
